Accept Certificate Type CLI Arguemtn (#1)
This tool supports the notion of linting different types of certificates (leafs, intermediates, and roots), however it does not accept the certificate type as an argument to the tool and thus always defaults to `subscriber` which leads to incorrect linting paths to be executed for intermediates and CAs.

This change updates `main` to parse out a second CLI argument signifying the type of certificate being linted. In order to maintain currently expected behavior, the default type chosen is `subscriber`.